requirements.txt
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pygame==2.0.1
argparse
